整理代码,添加几个宏简化代码

+++ Learn_opengl/Time.cpp
@@ -1,9 +1,9 @@
 #include "Time.h"
 
-float Time::timeScale = 1;
-float Time::time = 0;
-float Time::unscaledTime = 0;
-float Time::fixedTime = 0;
-float Time::deltaTime = 0;
-float Time::unscaledDeltaTime = 0;
-float Time::frameCount = 0;
+double Time::timeScale = 1;
+double Time::time = 0;
+double Time::unscaledTime = 0;
+double Time::fixedTime = 0;
+double Time::deltaTime = 0;
+double Time::unscaledDeltaTime = 0;
+double Time::frameCount = 0;
